day8/functions-signalr-v1/requirements.txt
azure-functions
openai>=1.40.0
requests
python-dotenv
